% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSigData.R
\docType{data}
\name{preCal.data}
\alias{preCal.data}
\title{pre-calculated penalization factors and molecular concepts coefficients}
\format{A list of characters. Each element contains information for a single <br>
gene in a tab delimited format. After splitting, the first <br>
element is Entrez Gene ID, the second one is a penalization factor <br>
used in uniConSig, the rest are the molecular concepts which are <br>
related to this gene and their coefficients.}
\usage{
data(preCal.data)
}
\description{
This file provides the coefficients for uniConSig calculation. <br> The details of the coefficients can be found in the package "uniConSig"
}
\details{
uniConSigData.
}
\examples{
library(uniConSigData)
data(preCal.data)
length(preCal.data)
class(preCal.data)
preCal.data[[1]]
}
\author{
Xu Chi <xuchi_chisquare@hotmail.com>
}
\keyword{datasets}
